+++ node/pkg/common/readiness.go
@@ -22,6 +22,7 @@ const (
 	ReadinessNeonSyncing       readiness.Component = "neonSyncing"
 	ReadinessTerra2Syncing     readiness.Component = "terra2Syncing"
 	ReadinessInjectiveSyncing  readiness.Component = "injectiveSyncing"
+	ReadinessPythNetSyncing    readiness.Component = "pythnetSyncing"
 
 	ReadinessMetaOSSyncing readiness.Component = "metaOSSyncing"
 )