+++ sdk/js/src/utils/array.ts
@@ -6,20 +6,27 @@ import {
   nativeStringToHexAlgorand,
   uint8ArrayToNativeStringAlgorand,
 } from "../algorand";
+import { canonicalAddress, humanAddress } from "../cosmos";
 import { buildTokenId } from "../cosmwasm/address";
-import { canonicalAddress, humanAddress, isNativeDenom } from "../terra";
+import { isNativeDenom } from "../terra";
 import {
   ChainId,
   ChainName,
   CHAIN_ID_ALGORAND,
   CHAIN_ID_NEAR,
+  CHAIN_ID_INJECTIVE,
+  CHAIN_ID_OSMOSIS,
+  CHAIN_ID_SUI,
+  CHAIN_ID_APTOS,
   CHAIN_ID_SOLANA,
   CHAIN_ID_TERRA,
   CHAIN_ID_TERRA2,
+  CHAIN_ID_WORMHOLE_CHAIN,
   CHAIN_ID_UNSET,
   coalesceChainId,
   isEVMChain,
   isTerraChain,
+  CHAIN_ID_PYTHNET,
 } from "./consts";
 
 /**
@@ -68,7 +75,7 @@ export const tryUint8ArrayToNative = (
   const chainId = coalesceChainId(chain);
   if (isEVMChain(chainId)) {
     return hexZeroPad(hexValue(a), 20);
-  } else if (chainId === CHAIN_ID_SOLANA) {
+  } else if (chainId === CHAIN_ID_SOLANA || chainId === CHAIN_ID_PYTHNET) {
     return new PublicKey(a).toString();
   } else if (isTerraChain(chainId)) {
     const h = uint8ArrayToHex(a);
@@ -77,14 +84,25 @@ export const tryUint8ArrayToNative = (
     } else {
       if (chainId === CHAIN_ID_TERRA2 && !isLikely20ByteTerra(h)) {
         // terra 2 has 32 byte addresses for contracts and 20 for wallets
-        return humanAddress(a);
+        return humanAddress("terra", a);
       }
-      return humanAddress(a.slice(-20));
+      return humanAddress("terra", a.slice(-20));
     }
   } else if (chainId === CHAIN_ID_ALGORAND) {
     return uint8ArrayToNativeStringAlgorand(a);
+  } else if (chainId == CHAIN_ID_WORMHOLE_CHAIN) {
+    // wormhole-chain addresses are always 20 bytes.
+    return humanAddress("wormhole", a.slice(-20));
   } else if (chainId === CHAIN_ID_NEAR) {
     throw Error("uint8ArrayToNative: Near not supported yet.");
+  } else if (chainId === CHAIN_ID_INJECTIVE) {
+    throw Error("uint8ArrayToNative: Injective not supported yet.");
+  } else if (chainId === CHAIN_ID_OSMOSIS) {
+    throw Error("uint8ArrayToNative: Osmosis not supported yet.");
+  } else if (chainId === CHAIN_ID_SUI) {
+    throw Error("uint8ArrayToNative: Sui not supported yet.");
+  } else if (chainId === CHAIN_ID_APTOS) {
+    throw Error("uint8ArrayToNative: Aptos not supported yet.");
   } else if (chainId === CHAIN_ID_UNSET) {
     throw Error("uint8ArrayToNative: Chain id unset");
   } else {
@@ -176,7 +194,7 @@ export const tryNativeToHexString = (
   const chainId = coalesceChainId(chain);
   if (isEVMChain(chainId)) {
     return uint8ArrayToHex(zeroPad(arrayify(address), 32));
-  } else if (chainId === CHAIN_ID_SOLANA) {
+  } else if (chainId === CHAIN_ID_SOLANA || chainId === CHAIN_ID_PYTHNET) {
     return uint8ArrayToHex(zeroPad(new PublicKey(address).toBytes(), 32));
   } else if (chainId === CHAIN_ID_TERRA) {
     if (isNativeDenom(address)) {
@@ -193,8 +211,18 @@ export const tryNativeToHexString = (
     return buildTokenId(address);
   } else if (chainId === CHAIN_ID_ALGORAND) {
     return nativeStringToHexAlgorand(address);
+  } else if (chainId == CHAIN_ID_WORMHOLE_CHAIN) {
+    return uint8ArrayToHex(zeroPad(canonicalAddress(address), 32));
   } else if (chainId === CHAIN_ID_NEAR) {
     throw Error("hexToNativeString: Near not supported yet.");
+  } else if (chainId === CHAIN_ID_INJECTIVE) {
+    throw Error("hexToNativeString: Injective not supported yet.");
+  } else if (chainId === CHAIN_ID_OSMOSIS) {
+    throw Error("hexToNativeString: Osmosis not supported yet.");
+  } else if (chainId === CHAIN_ID_SUI) {
+    throw Error("hexToNativeString: Sui not supported yet.");
+  } else if (chainId === CHAIN_ID_APTOS) {
+    throw Error("hexToNativeString: Aptos not supported yet.");
   } else if (chainId === CHAIN_ID_UNSET) {
     throw Error("hexToNativeString: Chain id unset");
   } else {