+++ sdk/js/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## 0.6.0
+
+### Added
+
+Wormhole chain devnet support
+
+human-readable part parameter to `humanAddress` function
+
+### Changed
+
+`canonicalAddress` and `humanAddress` functions moved from terra to cosmos module
+
+## 0.5.2
+
+### Added
+
+Support for PythNet
+Chain ids for Arbitrum, Optimism, and Gnosis
+
+## 0.5.1
+
+### Added
+
+Chain ids for Injective, Osmosis, Sui, and Aptos
+
+## 0.5.0
+
+### Changed
+
+Use `@certusone/wormhole-sdk-proto-web` and `@certusone/wormhole-sdk-wasm` packages
+
 ## 0.4.5
 
 ### Changed